fix: answer the construct-split review — restore CI coverage, inlining, and the public header
Round 1 found two minors and two nits, all taken:

  - tasks.py's C_SOURCES and C_TRANSLATION_UNITS globs were
    non-recursive, so the folder modules dropped out of CI's format and
    analyze legs; both are now rglob, which also restores the coverage
    of src/meta/* that #68 silently dropped the same way
  - bind_positional, bind_keywords, and fill_defaults returned to
    construct.c, their only caller TU: they are static again, the
    cross-TU calls are gone, and the benchmark reads 10.3 us/type and
    52.3 inst ns against the pre-split 10.5 and 54.6 -- within the
    51.6-54.6 band three runs measured today
  - the folder header forward-declares StructType instead of including
    types.h, per the #68 recipe, and each construct TU includes types.h
    itself
  - src/construct.h is byte-identical to its pre-split self again: the
    five public declarations only; the folder header now carries just
    the cross-file internals (field_lookup, find_field, write_slot),
    which are the only two helpers left on the shared surface

The gate ran unscoped -- the whole tree, exactly what CI runs -- and is
green across 3.10-3.15, free-threaded, the in-file C tests, and the
analyzers, now genuinely covering the folder files.
